+++ src/components/common/buttons/ScanQRButton.js
