+++ src/components/common/buttons/__tests__/CopyButton.js
