+++ src/components/common/dialogs/EventDialog.js
@@ -2,10 +2,10 @@
 import React from 'react'
 import { StyleSheet, View } from 'react-native'
 import { Dialog, Paragraph, Portal, Text } from 'react-native-paper'
-import type { TransactionEvent } from '../../lib/gundb/UserStorage'
-import { Avatar, BigGoodDollar } from '../common'
-import CustomButton from './CustomButton'
-import Section from './Section'
+import type { TransactionEvent } from '../../../lib/gundb/UserStorage'
+import { Avatar, BigGoodDollar } from '..'
+import CustomButton from '../buttons/CustomButton'
+import Section from '../layout/Section'
 
 export type EventDialogProps = {
   visible: boolean,
