+++ src/components/common/dialogs/__tests__/EventDialog.js
@@ -3,14 +3,19 @@ import React from 'react'
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer'
 
-import CustomDialog from '../CustomDialog'
-import { Wrapper } from '../index'
+import EventDialog from '../EventDialog'
+import { Wrapper } from '../../index'
+import { mockEvent } from '../../../dashboard/__tests__/__util__'
 
-describe('CustomDialog', () => {
+const event = mockEvent('Received')
+
+describe('EventDialog', () => {
   it('renders without errors', () => {
     const tree = renderer.create(
       <Wrapper>
-        <CustomDialog>Next</CustomDialog>
+        <EventDialog visible event={event}>
+          Next
+        </EventDialog>
       </Wrapper>
     )
     expect(tree.toJSON()).toBeTruthy()
@@ -19,7 +24,9 @@ describe('CustomDialog', () => {
   it('matches snapshot', () => {
     const component = renderer.create(
       <Wrapper>
-        <CustomDialog>Next</CustomDialog>
+        <EventDialog visible event={event}>
+          Next
+        </EventDialog>
       </Wrapper>
     )
     const tree = component.toJSON()
