+++ src/components/common/form/InputGoodDollar.js
@@ -2,7 +2,7 @@
 
 import React from 'react'
 import { TextInput } from 'react-native-paper'
-import { maskToWei, weiToMask } from '../../lib/wallet/utils'
+import { maskToWei, weiToMask } from '../../../lib/wallet/utils'
 
 type Props = {
   onChangeWei: number => void,
