+++ src/components/common/form/__tests__/InputGoodDollar.js
