+++ src/components/common/layout/Section.js
@@ -2,7 +2,7 @@
 import React, { Component } from 'react'
 import { StyleSheet, Text, View } from 'react-native'
 import normalize from 'react-native-elements/src/helpers/normalizeText'
-import { fontStyle } from './styles'
+import { fontStyle } from '../styles'
 
 const getFlexStylesFromProps = props => {
   const { justifyContent, alignItems, grow } = props
