+++ src/components/common/layout/__tests__/Section.js
