+++ src/components/common/view/Avatar.js
