+++ src/components/common/view/__tests__/Avatar.js
