+++ src/components/common/view/__tests__/BigGoodDollar.js
