+++ src/components/common/view/__tests__/BigNumber.js
