+++ src/components/common/view/__tests__/LoadingIndicator.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import GDStore from '../../../lib/undux/GDStore'
+import GDStore from '../../../../lib/undux/GDStore'
 import LoadingIndicator from '../LoadingIndicator'
 
 // import LoadingIndicator, { setLoadingWithStore } from '../LoadingIndicator'
