+++ src/components/common/view/__tests__/NumPadKeyboard.js
