+++ src/components/common/view/__tests__/UserAvatar.js
