+++ src/components/dashboard/FeedItems/ListEventItem.js
@@ -2,7 +2,7 @@
 import React from 'react'
 import { Avatar } from 'react-native-paper'
 import { Text, View } from 'react-native-web'
-import BigGoodDollar from '../../common/BigGoodDollar'
+import { BigGoodDollar } from '../../common/'
 import { getFormattedDateTime } from '../../../lib/utils/FormatDate'
 import { listStyles } from './EventStyles'
 import type { FeedEventProps } from './EventProps'
