+++ src/components/dashboard/Send.js
@@ -5,7 +5,7 @@ import userStorage from '../../lib/gundb/UserStorage'
 import logger from '../../lib/logger/pino-logger'
 import { readCode } from '../../lib/share'
 import { useDialog } from '../../lib/undux/utils/dialog'
-import InputRecipient from '../../components/common/InputRecipient'
+import InputRecipient from '../common/form/InputRecipient'
 import isMobilePhone from '../../lib/validators/isMobilePhone'
 import goodWallet from '../../lib/wallet/GoodWallet'
 import { CustomButton, IconButton, Section, TopBar, Wrapper } from '../common'
